5.	Create a simple pipeline which would do the following actions:
a.	Increase the Patch part of the jar version ( 1.0.0 - > 1.0.1-> …) automatically
b.	Compile the code
c.	Package it into an artifact.
d.	Create an artifact item for the build.
e.	Create a docker image containing the artifact  - use Dockerfile:
i.	Tag the Docker image as the Jar version automatically.
ii.	The Docker image shouldn't run with root.
f.	Push the docker image that was created in the previous step to Docker Hub
g.	Download and run the docker image.
6.	Once it’s running, we will come and pick you up immediately.